@@ -5,18 +5,18 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=5.5.9",
-        "laravel/framework": "5.1.*",
+        "php": "^7.4|^8.0",
+        "laravel/framework": "8.*",
         "guzzlehttp/guzzle": "~6.0",
-        "freddiedfre/africas_talking_laravel_5": "dev-master",
         "twilio/sdk": "^5.13",
-        "doctrine/dbal": "^2.5"
+        "doctrine/dbal": "^2.5",
+        "africastalking/africastalking": "^3.0"
     },
     "require-dev": {
-        "fzaninotto/faker": "~1.4",
-        "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~4.0",
-        "phpspec/phpspec": "~2.1"
+        "fakerphp/faker": "~1.12.x-dev",
+        "mockery/mockery": "1.5.0",
+        "phpunit/phpunit": "~9.5.13",
+        "phpspec/phpspec": "~8.0.x-dev"
     },
     "autoload": {
         "classmap": [
