@@ -20,14 +20,14 @@ class RouteServiceProvider extends ServiceProvider {
     /**
      * Define your route model bindings, pattern filters, etc.
      *
-     * @param  \Illuminate\Routing\Router  $router
      * @return void
      */
-    public function boot(Router $router) {
+    public function boot() {
         //
 
-        parent::boot($router);
+        parent::boot();
 
+        $router = $this;
         $router->bind('broker', function($value) {
             try {
                 return \App\Models\User::withStatus()->findOrFail($value);
